pyserver/requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
flask_swagger==0.2.14
flask_swagger_ui==4.11.1
modules==1.0.0
psutil==5.9.5
python-telegram-bot==20.4
Werkzeug==2.3.6
